# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: BuilderConnect API
  description: |
    Smart Construction Marketplace & Project Lifecycle Management Platform API.

    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Roles
    - CLIENT: Project owners who post projects
    - BUILDER: Construction professionals who bid on projects
    - SUPPLIER: Material suppliers
    - SUPERVISOR: On-site supervisors
    - INSPECTOR: Quality inspectors
    - SUPPORT_AGENT: Customer support
    - ADMIN: Platform administrators
    - SUPER_ADMIN: Full system access

  version: 2.0.0
  contact:
    name: BuilderConnect Support
    email: support@builderconnect.pk
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /client/projects:
    get:
      tags:
        - Projects
      summary: Get client's projects
      operationId: getClientProjects
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'
    post:
      tags:
        - Projects
      summary: Create a new project
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
  /client/projects/{id}:
    get:
      tags:
        - Projects
      summary: Get project details
      operationId: getClientProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
  /client/projects/{id}/publish:
    post:
      tags:
        - Projects
      summary: Publish a draft project
      operationId: publishProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
  /client/projects/{projectId}/award/{bidId}:
    post:
      tags:
        - Projects
      summary: Award project to a bidder
      operationId: awardProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project awarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
  /projects:
    get:
      tags:
        - Projects
      summary: Search open projects (marketplace)
      operationId: searchProjects
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
        - name: minBudget
          in: query
          schema:
            type: number
        - name: maxBudget
          in: query
          schema:
            type: number
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'

  # ==================== BIDS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
  schemas:
...
    ProjectCreateRequest:
      type: object
      required:
        - title
        - description
        - categoryId
        - city
        - budgetMin
        - budgetMax
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
          format: int64
        city:
          type: string
        locationAddress:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        deadline:
          type: string
          format: date
        isUrgent:
          type: boolean
        requiresInspection:
          type: boolean
...
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectNumber:
          type: string
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        city:
          type: string
        locationAddress:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        deadline:
          type: string
          format: date
        status:
          type: string
        isUrgent:
          type: boolean
        requiresInspection:
          type: boolean
        bidCount:
          type: integer
        clientName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ProjectStatus:
      type: string
      enum:
        - DRAFT
        - OPEN
        - BIDDING
        - AWARDED
        - IN_PROGRESS
        - ON_HOLD
        - COMPLETED
        - CANCELLED
        - DISPUTED
...
    ProjectPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
...